feat: wip climate data parsing scripts
