+++ apps/picsa-server/scripts/climate-data-parser/.gitignore
@@ -1,0 +1,3 @@
+output
+input/*
+!input/*.gitkeep
